+++ libMx/d_soft_motor.h
@@ -8,7 +8,7 @@
  *
  *------------------------------------------------------------------------
  *
- * Copyright 1999, 2001-2002, 2004 Illinois Institute of Technology
+ * Copyright 1999, 2001-2002, 2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -47,7 +47,7 @@ MX_API mx_status_type mxd_soft_motor_positive_limit_hit( MX_MOTOR *motor );
 MX_API mx_status_type mxd_soft_motor_negative_limit_hit( MX_MOTOR *motor );
 MX_API mx_status_type mxd_soft_motor_get_parameter( MX_MOTOR *motor );
 MX_API mx_status_type mxd_soft_motor_set_parameter( MX_MOTOR *motor );
-MX_API mx_status_type mxd_soft_motor_simultaneous_start( int num_motor_records,
+MX_API mx_status_type mxd_soft_motor_simultaneous_start( long num_motor_records,
 						MX_RECORD **motor_record_array,
 						double *position_array,
 						int flags );
