+++ libMx/d_spec_motor.h
@@ -6,7 +6,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 2004 Illinois Institute of Technology
+ * Copyright 2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -52,7 +52,7 @@ MX_API mx_status_type mxd_spec_motor_negative_limit_hit( MX_MOTOR *motor );
 MX_API mx_status_type mxd_spec_motor_find_home_position( MX_MOTOR *motor );
 MX_API mx_status_type mxd_spec_motor_get_parameter( MX_MOTOR *motor );
 MX_API mx_status_type mxd_spec_motor_set_parameter( MX_MOTOR *motor );
-MX_API mx_status_type mxd_spec_motor_simultaneous_start( int num_motor_records,
+MX_API mx_status_type mxd_spec_motor_simultaneous_start( long num_motor_records,
 						MX_RECORD **motor_record_array,
 						double *position_array,
 						int flags );
