+++ libMx/i_network_gpib.h
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 2005 Illinois Institute of Technology
+ * Copyright 2005-2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -59,35 +59,35 @@ MX_API mx_status_type mxi_network_gpib_close( MX_RECORD *record );
 MX_API mx_status_type mxi_network_gpib_resynchronize( MX_RECORD *record );
 
 MX_API mx_status_type mxi_network_gpib_open_device( MX_GPIB *gpib,
-							int address );
+							long address );
 MX_API mx_status_type mxi_network_gpib_close_device( MX_GPIB *gpib,
-							int address );
+							long address );
 MX_API mx_status_type mxi_network_gpib_read( MX_GPIB *gpib,
-						int address,
+						long address,
 						char *buffer,
 						size_t max_bytes_to_read,
 						size_t *bytes_read,
-						int flags );
+						unsigned long flags );
 MX_API mx_status_type mxi_network_gpib_write( MX_GPIB *gpib,
-						int address,
+						long address,
 						char *buffer,
 						size_t bytes_to_write,
 						size_t *bytes_written,
-						int flags );
+						unsigned long flags );
 MX_API mx_status_type mxi_network_gpib_interface_clear( MX_GPIB *gpib );
 MX_API mx_status_type mxi_network_gpib_device_clear( MX_GPIB *gpib );
 MX_API mx_status_type mxi_network_gpib_selective_device_clear( MX_GPIB *gpib,
-								int address );
+								long address );
 MX_API mx_status_type mxi_network_gpib_local_lockout( MX_GPIB *gpib );
 MX_API mx_status_type mxi_network_gpib_remote_enable( MX_GPIB *gpib,
-								int address );
+								long address );
 MX_API mx_status_type mxi_network_gpib_go_to_local( MX_GPIB *gpib,
-								int address );
-MX_API mx_status_type mxi_network_gpib_trigger( MX_GPIB *gpib, int address );
+								long address );
+MX_API mx_status_type mxi_network_gpib_trigger( MX_GPIB *gpib, long address );
 MX_API mx_status_type mxi_network_gpib_wait_for_service_request( MX_GPIB *gpib,
 							double timeout );
 MX_API mx_status_type mxi_network_gpib_serial_poll( MX_GPIB *gpib,
-					int address,
+					long address,
 					unsigned char *serial_poll_byte );
 MX_API mx_status_type mxi_network_gpib_serial_poll_disable( MX_GPIB *gpib );
